+++ actions/docs/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added a `Table` data model.
 - It's now possible to use python types such as list/dict directly as action parameters/return type without requiring a pydantic model.
+- When linting, if a parameter is a `DataSource` it's not required to have a docstring.
+  - Note: as it works just on the AST, it currently just checks if the python name ends with `DataSource` (instead of doing full type inference).
 
 ## 1.1.4 - 2024-11-26